rename portafolio to wallet and make OperationEmptyState global
