+++ src/app/(protected)/operations/components/OperationsEmptyState.tsx
